- complete functionality in search page with pagination.

+++ app/src/main/java/com/speertech/testapp/model/SearchResults.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class SearchResults(
     @SerializedName("total_count")
-    val totalCount: Long,
+    val totalCount: Int,
     @SerializedName("incomplete_results")
     val incompleteResults: Boolean,
     @SerializedName("items")
